Bug fix for deleting account when account is not empty

+++ src/main/java/seedu/duke/AccountList.java
@@ -14,7 +14,7 @@ public class AccountList {
     // Hashmap of ArrayList of account in the future and randomly generate an ID for that account
     private final HashMap<Currency, Account> accountHashMap;
 
-    public AccountList () {
+    public AccountList() {
         accountHashMap = new HashMap<>();
     }
 
@@ -25,7 +25,7 @@ public class AccountList {
      * @param initialBalance The initial balance of the new account to be added
      * @throws AccountAlreadyExistsException If the account already exists
      */
-    public void addAccount (Currency currency, float initialBalance) throws AccountAlreadyExistsException {
+    public void addAccount(Currency currency, float initialBalance) throws AccountAlreadyExistsException {
         if (accountHashMap.containsKey(currency)) {
             throw new AccountAlreadyExistsException();
         }
@@ -38,11 +38,11 @@ public class AccountList {
      * @param currency The currency type of the account to be deleted
      * @throws NoAccountException If the account does not exist
      */
-    public void deleteAccount (Currency currency) throws NoAccountException, AccountNotEmptyException {
+    public void deleteAccount(Currency currency) throws NoAccountException, AccountNotEmptyException {
         if (!accountHashMap.containsKey(currency)) {
             throw new NoAccountException();
         }
-        if ((int) accountHashMap.get(currency).getBalance() != 0) {
+        if (accountHashMap.get(currency).getBalance() != (float) 0) {
             throw new AccountNotEmptyException();
         }
         accountHashMap.remove(currency);
@@ -53,7 +53,7 @@ public class AccountList {
      *
      * @return An array list containing all the accounts.
      */
-    public ArrayList<Account> getAllAccounts () {
+    public ArrayList<Account> getAllAccounts() {
         ArrayList<Account> accounts = new ArrayList<>();
         accountHashMap.forEach(((currency, account) -> {
             accounts.add(account);
@@ -68,7 +68,7 @@ public class AccountList {
      * @return the currency account
      * @throws NoAccountException if the user does not have an account for that currency
      */
-    public Account getAccount (Currency currency) throws NoAccountException {
+    public Account getAccount(Currency currency) throws NoAccountException {
         if (!accountHashMap.containsKey(currency)) {
             throw new NoAccountException();
         }
